require db 讀不到資料
