+++ backend/models/schInfoModel.js
@@ -1,26 +1,26 @@
-let mysql =require('mysql');
-let conn = mysql.createConnection({
-    host:'127.0.0.1',
-    user:'root',
-    password:'',
-    database:'chill_around'
-})
+// let mysql =require('mysql');
+// let conn = mysql.createConnection({
+//     host:'127.0.0.1',
+//     user:'root',
+//     password:'',
+//     database:'chill_around'
+// })
 
 
-conn.connect(function (err) {
+// conn.connect(function (err) {
 
-    if(err){
-        console.log('tell you:資料庫連線 異常');
-        console.log('-----------start---------------');
-        console.log(err.sqlMessage);
-        console.log('-----------end---------------');
+//     if(err){
+//         console.log('tell you:資料庫連線 異常');
+//         console.log('-----------start---------------');
+//         console.log(err.sqlMessage);
+//         console.log('-----------end---------------');
         
-    }else{
-        console.log('tell you:資料庫連線 sch正常');
+//     }else{
+//         console.log('tell you:資料庫連線 sch正常');
 
-    }
+//     }
     
-});
+// });
 
 
 // hero img sql
